video: ssd2828: Use MIPI DCS commands to retrieve the LCD panel id
Or at least try to do this. With the possibility of bi-directional
communication, it seems natural to expect that MIPI displays
should be able to handle automatic configuration for the screen
resolution and timings nicely. But the reality is not so pretty.

It appears that the manufacturers of MIPI LCD panels are not very
disciplined and do not follow any kind of standard guidelines
when it comes to providing the panel identification information.
MIPI DSI even has two alternative standard DCS commands for
potentially reading the vendor/panel id: MIPI_DCS_GET_DISPLAY_ID
and MIPI_DCS_READ_DDB_START. But these commands do not seem to
be widely implemented in real hardware. Moreover, the vendors
seem to sometimes invent their own custom new commands, while
ignoring the standard ones.

This particular patch just probes both of the standard DCS
commands and also a few non-standard ones. With the hope to
actually extract anything unique and/or meaningful from the
LCD panel. This information is then printed to the u-boot
console.

We don't know what the future brings. Maybe the manufacturers
are going to start acting responsibly some day. And will
implement proper panel identification.

Regarding what we have right now. The information retrieval
DCS commands still work if reducing the clock speed to the very
minimum and enabling only one lane on my MSI Primo81 tablet.
Potentially it means that one could perhaps swap two different
MIPI LCD panels (if they have a compatible connector) and still
have the software handling this fine by doing runtime selection
of the right settings (based on checking the panel id).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>