video: Add support for SSD2828 (parallel LCD to MIPI bridge)
SSD2828 can take pixel data coming from a parallel LCD interface
and translate it on the flight into MIPI DSI interface for
driving a MIPI compatible TFT display. SSD2828 is configured
over SPI interface, which may or may not have MISO pin wired
up on some hardware. So a write-only SPI mode also has to be
supported.

The SSD2828 support code is implemented as a utility function
and needs to be called from real display drivers, which are
responsible for driving parallel LCD hardware in front of the
video pipeline. The usage instructions are provided as comments
in the header file.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>

v2: Fixed missing static keyword for one function and a bug in
    'mipi_dsi_loosely_packed_pixel_format' handling logic