sunxi: Use more realistic size limit for FEL SPL
The older 16 KiB limit was overly optimistic. The FEL SPL binary
is loaded at the address 0x2000 (which is also the entry point),
and the stack pointer is set by the FEL BROM code to 0x5E00 on
Allwinner A20 right at the start. This gives us around 15872
bytes for everything (code, data and stack). Considering that
the stack usage is somewhere between 1 KiB and 1.5 KiB (mostly
because of printf buffers), setting 14 KiB as the FEL SPL size
limit is reasonable.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>