HACK: SPI NOR Flash emulation demo (uses SPI2 on Allwinner A13)
It is possible to emulate the Orange Pi PC firmware in the SPI NOR
flash by running a custom U-Boot on an Allwinner A13 board.

The SPI2 pins on Allwinner A13 board (A13-OLinuXino-Micro) are connected
to the SPI0 pins on the Orange Pi PC board. Then the Orange Pi PC is
powered on without a bootable SD card in the SD card slot and is able
to load the SPL over SPI (unfortunately the SD card appears to have
a higher priority in the boot sequence than the SPI NOR Flash). The
UART serial console output looks like this:

== Serial console on https://linux-sunxi.org/Olimex_A13-OLinuXino-Micro

   U-Boot 2016.03-dirty (Jan 01 1970 - 00:00:00 +0000) Allwinner Technology

   CPU:   Allwinner A13 (SUN5I)
   I2C:   ready
   DRAM:  256 MiB

   Emulating SPI NOR flash on SPI2...
   The emulated NOR flash has been fully read, stopping.

   Emulating SPI NOR flash on SPI2...
   The emulated NOR flash has been fully read, stopping.

== Serial console on https://linux-sunxi.org/Xunlong_Orange_Pi_PC

   U-Boot SPL 2016.01-rc2-dirty (Jan 01 1970 - 00:00:00)
   DRAM: 1024 MiB
   Card did not respond to voltage select!
   Could not determine boot source

   resetting ...

   U-Boot SPL 2016.01-rc2-dirty (Jan 01 1970 - 00:00:00)
   DRAM: 1024 MiB
   Card did not respond to voltage select!
   Could not determine boot source

   resetting ...

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>