HACK: SPI NOR Flash emulation demo code (uses SPI2 on Allwinner A13)
It is possible to do a simplistic emulation of SPI NOR Flash by
running a patched U-Boot on an Allwinner A13 board.

The Allwinner's BROM tries to read the first stage bootloader from
SPI0 as one of the boot methods, typically after everything else
has failed. It first requests 256 bytes starting from the address 0,
verifies correctness of the header and then rewinds back to the
address 0 and starts requesting 2048 byte blocks one after another.
We just produce this particular output and the BROM seems to be happy.

This particular patch just includes 8K of the fel-sdboot.sunxi
binary as a generic payload. It is not very useful for checking
whether the boot was successful, because the BROM normally
fallbacks to the FEL mode in the case of error anyway. But this
payload can be replaced with the U-Boot SPL or the Allwinner's
boot0 suitable for the target hardware.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>